Docker/Docker_Compose/app/requirements.txt
Flask==2.3.2
mysql-connector-python==8.2.0
redis==5.0.1
